chore: refactor format output adjustments
I had previously done this in an overly complicated manner. This PR cleans
it up a bit.
