+++ src/extractors/root-extractor.test.js
@@ -32,73 +32,6 @@ describe('RootExtractor', () => {
 
     assert.equal(url, null);
   });
-  it('returns text content if text is passed as contentType', () => {
-    const fullUrl =
-      'http://nymag.com/daily/intelligencer/2016/09/trump-discussed-usd25k-donation-with-florida-ag-not-fraud.html';
-    const html = fs.readFileSync(
-      './src/extractors/custom/nymag.com/fixtures/test.html',
-      'utf8'
-    );
-    const $ = cheerio.load(html);
-
-    const { content } = RootExtractor.extract(NYMagExtractor, {
-      url: fullUrl,
-      html,
-      $,
-      metaCache: [],
-      fallback: false,
-      contentType: 'text',
-    });
-
-    const htmlRe = /<[a-z][\s\S]*>/g;
-
-    assert.equal(htmlRe.test(content), false);
-  });
-  it('returns markdown if markdown is passed as contentType', () => {
-    const fullUrl =
-      'http://nymag.com/daily/intelligencer/2016/09/trump-discussed-usd25k-donation-with-florida-ag-not-fraud.html';
-    const html = fs.readFileSync(
-      './src/extractors/custom/nymag.com/fixtures/test.html',
-      'utf8'
-    );
-    const $ = cheerio.load(html);
-
-    const { content } = RootExtractor.extract(NYMagExtractor, {
-      url: fullUrl,
-      html,
-      $,
-      metaCache: [],
-      fallback: false,
-      contentType: 'markdown',
-    });
-
-    const htmlRe = /<[a-z][\s\S]*>/;
-    const markdownRe = /\[[\w\s]+\]\(.*\)/;
-
-    assert.equal(htmlRe.test(content), false);
-    assert.equal(markdownRe.test(content), true);
-  });
-  it('also can select type on Generic Extractor', () => {
-    const fullUrl =
-      'http://www.vulture.com/2016/08/dc-comics-greg-berlanti-c-v-r.html';
-
-    const html = fs.readFileSync('./fixtures/vulture.html', 'utf-8');
-    const $ = cheerio.load(html);
-    const { content } = RootExtractor.extract(undefined, {
-      url: fullUrl,
-      html,
-      $,
-      metaCache: [],
-      fallback: false,
-      contentType: 'markdown',
-    });
-
-    const htmlRe = /<[a-z][\s\S]*>/;
-    const markdownRe = /\[[\w\s]+\]\(.*\)/;
-
-    assert.equal(htmlRe.test(content), false);
-    assert.equal(markdownRe.test(content), true);
-  });
 });
 
 describe('cleanBySelectors($content, $, { clean })', () => {
